Xcode 12.2

+++ Managed Objects/TSSTManagedGroup.m
@@ -513,7 +513,7 @@
     NSPDFImageRep * rep = [self instance];
     TSSTPage * imageDescription;
     NSMutableSet * pageSet = [NSMutableSet set];
-    int imageCount = [rep pageCount];
+    NSInteger imageCount = [rep pageCount];
     int pageNumber;
     for (pageNumber = 0; pageNumber < imageCount; ++pageNumber)
     {
